Rename to MacrosEvolution Fixes #278
