+++ app/profile/TopBar.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import UserSelector from '../../components/UserSelector'
 
-export default function TopBar() {
+export function TopBar() {
   return (
     <div className={`flex justify-between bg-slate-600 p-4`}>
       <Link href="/" className={`my-auto text-xl`}>
